requirements.txt
alembic==1.15.2
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fuzzywuzzy==0.18.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.1
Mako==1.3.9
MarkupSafe==3.0.2
python-dotenv==1.1.0
python-Levenshtein==0.27.1
RapidFuzz==3.12.2
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
